try to add pointwise lauch logic in c++

+++ include/flag_gems/utils.h
@@ -25,7 +25,7 @@ std::tuple<at::Tensor, int64_t, int64_t> permute_reduction_axes_right(const at::
 std::tuple<at::Tensor, int64_t, int64_t> permute_reduction_axes_right(
     const at::Tensor &tensor, at::OptionalIntArrayRef reduction_axes_opt);
 std::tuple<int64_t, int64_t, int64_t> parse_reduction_axes(const at::Tensor &tensor, int reduction_axis);
-int cdiv(int a, int b);
+int64_t cdiv(int64_t a, int64_t b);
 }  // namespace flag_gems::utils
 
 // Public declaration for reuse (implementation in lib/exponential_.cpp)
